+++ include/config.h
@@ -4,6 +4,7 @@
 #include <QString>
 #include <QObject>
 #include <QSize>
+#include <QByteArrayList>
 
 enum MapSortOrder {
     Group   =  0,
@@ -34,19 +35,27 @@ public:
         this->prettyCursors = true;
         this->collisionOpacity = 50;
         this->regionMapDimensions = QSize(32, 20);
+        this->showPlayerView = false;
+        this->showCursorTile = true;
     }
     void setRecentProject(QString project);
     void setRecentMap(QString map);
     void setMapSortOrder(MapSortOrder order);
     void setPrettyCursors(bool enabled);
+    void setGeometry(QByteArray, QByteArray, QByteArray, QByteArray, QByteArray);
     void setCollisionOpacity(int opacity);
     void setRegionMapDimensions(int width, int height);
+    void setShowPlayerView(bool enabled);
+    void setShowCursorTile(bool enabled);
     QString getRecentProject();
     QString getRecentMap();
     MapSortOrder getMapSortOrder();
     bool getPrettyCursors();
+    QMap<QString, QByteArray> getGeometry();
     int getCollisionOpacity();
     QSize getRegionMapDimensions();
+    bool getShowPlayerView();
+    bool getShowCursorTile();
 protected:
     QString getConfigFilepath();
     void parseConfigKeyValue(QString key, QString value);
@@ -55,16 +64,26 @@ protected:
 private:
     QString recentProject;
     QString recentMap;
+    QString stringFromByteArray(QByteArray);
+    QByteArray bytesFromString(QString);
     MapSortOrder mapSortOrder;
     bool prettyCursors;
+    QByteArray windowGeometry;
+    QByteArray windowState;
+    QByteArray mapSplitterState;
+    QByteArray eventsSlpitterState;
+    QByteArray mainSplitterState;
     int collisionOpacity;
     QSize regionMapDimensions;
+    bool showPlayerView;
+    bool showCursorTile;
 };
 
 extern PorymapConfig porymapConfig;
 
 enum BaseGameVersion {
     pokeruby,
+    pokefirered,
     pokeemerald,
 };