+++ include/mainwindow.h
@@ -17,6 +17,7 @@
 #include "tileseteditor.h"
 #include "regionmapeditor.h"
 #include "filterchildrenproxymodel.h"
+#include "newmappopup.h"
 
 namespace Ui {
 class MainWindow;
@@ -48,7 +49,10 @@ private slots:
     void onMapChanged(Map *map);
     void onMapNeedsRedrawing();
     void onTilesetsSaved(QString, QString);
+    void openNewMapPopupWindow(int, QVariant);
+    void onNewMapCreated();
 
+    void on_action_NewMap_triggered();
     void on_action_Save_triggered();
     void on_tabWidget_2_currentChanged(int index);
     void on_action_Exit_triggered();
@@ -71,6 +75,8 @@ private slots:
     void on_actionZoom_In_triggered();
     void on_actionZoom_Out_triggered();
     void on_actionBetter_Cursors_triggered();
+    void on_actionPlayer_View_Rectangle_triggered();
+    void on_actionCursor_Tile_Outline_triggered();
     void on_actionPencil_triggered();
     void on_actionPointer_triggered();
     void on_actionFlood_Fill_triggered();
@@ -99,6 +105,8 @@ private slots:
 
     void onOpenMapListContextMenu(const QPoint &point);
     void onAddNewMapToGroupClick(QAction* triggeredAction);
+    void onAddNewMapToAreaClick(QAction* triggeredAction);
+    void onAddNewMapToLayoutClick(QAction* triggeredAction);
     void onTilesetChanged(QString);
     void currentMetatilesSelectionChanged();
 
@@ -148,11 +156,18 @@ private slots:
 
     void on_actionRegion_Map_Editor_triggered();
 
+    void on_toolButton_ExpandAll_clicked();
+
+    void on_toolButton_CollapseAll_clicked();
+
+    void on_actionAbout_Porymap_triggered();
+
 private:
     Ui::MainWindow *ui;
     TilesetEditor *tilesetEditor = nullptr;
     RegionMapEditor *regionMapEditor = nullptr;
     FilterChildrenProxyModel *mapListProxyModel;
+    NewMapPopup *newmapprompt = nullptr;
     QStandardItemModel *mapListModel;
     QList<QStandardItem*> *mapGroupItemsList;
     QMap<QString, QModelIndex> mapListIndexes;
@@ -204,6 +219,7 @@ private:
     void initMapSortOrder();
     void setProjectSpecificUIVisibility();
     void loadUserSettings();
+    void restoreWindowState();
     bool openRecentProject();
     void updateTilesetEditor();
     QString getEventGroupFromTabWidget(QWidget *tab);