+++ src/editor.cpp
@@ -18,6 +18,8 @@ Editor::Editor(Ui::MainWindow* ui)
     this->ui = ui;
     this->selected_events = new QList<DraggablePixmapItem*>;
     this->settings = new Settings();
+    this->playerViewRect = new MovableRect(&this->settings->playerViewRectEnabled, 30 * 8, 20 * 8, qRgb(255, 255, 255));
+    this->cursorMapTileRect = new CursorTileRect(&this->settings->cursorTileRectEnabled, qRgb(255, 255, 255));
 }
 
 void Editor::saveProject() {
@@ -35,7 +37,7 @@ void Editor::save() {
 }
 
 void Editor::undo() {
-    if (current_view) {
+    if (current_view && map_item->paintingEnabled) {
         map->undo();
         map_item->draw();
         collision_item->draw();
@@ -43,7 +45,7 @@ void Editor::undo() {
 }
 
 void Editor::redo() {
-    if (current_view) {
+    if (current_view && map_item->paintingEnabled) {
         map->redo();
         map_item->draw();
         collision_item->draw();
@@ -60,10 +62,10 @@ void Editor::closeProject() {
 void Editor::setEditingMap() {
     current_view = map_item;
     if (map_item) {
+        map_item->paintingEnabled = true;
         displayMapConnections();
         map_item->draw();
         map_item->setVisible(true);
-        map_item->setEnabled(true);
         setConnectionsVisibility(ui->checkBox_ToggleBorder->isChecked());
     }
     if (collision_item) {
@@ -74,6 +76,8 @@ void Editor::setEditingMap() {
     }
     setBorderItemsVisible(ui->checkBox_ToggleBorder->isChecked());
     setConnectionItemsVisible(false);
+    this->cursorMapTileRect->stopSingleTileMode();
+    this->cursorMapTileRect->setVisibility(true);
 }
 
 void Editor::setEditingCollision() {
@@ -85,6 +89,7 @@ void Editor::setEditingCollision() {
         setConnectionsVisibility(ui->checkBox_ToggleBorder->isChecked());
     }
     if (map_item) {
+        map_item->paintingEnabled = true;
         map_item->setVisible(false);
     }
     if (events_group) {
@@ -92,6 +97,8 @@ void Editor::setEditingCollision() {
     }
     setBorderItemsVisible(ui->checkBox_ToggleBorder->isChecked());
     setConnectionItemsVisible(false);
+    this->cursorMapTileRect->setSingleTileMode();
+    this->cursorMapTileRect->setVisibility(true);
 }
 
 void Editor::setEditingObjects() {
@@ -100,8 +107,8 @@ void Editor::setEditingObjects() {
         events_group->setVisible(true);
     }
     if (map_item) {
+        map_item->paintingEnabled = false;
         map_item->setVisible(true);
-        map_item->setEnabled(false);
         setConnectionsVisibility(ui->checkBox_ToggleBorder->isChecked());
     }
     if (collision_item) {
@@ -109,14 +116,16 @@ void Editor::setEditingObjects() {
     }
     setBorderItemsVisible(ui->checkBox_ToggleBorder->isChecked());
     setConnectionItemsVisible(false);
+    this->cursorMapTileRect->setSingleTileMode();
+    this->cursorMapTileRect->setVisibility(false);
 }
 
 void Editor::setEditingConnections() {
     current_view = map_item;
     if (map_item) {
+        map_item->paintingEnabled = false;
         map_item->draw();
         map_item->setVisible(true);
-        map_item->setEnabled(false);
         populateConnectionMapPickers();
         ui->label_NumConnections->setText(QString::number(map->connections.length()));
         setConnectionsVisibility(false);
@@ -137,6 +146,8 @@ void Editor::setEditingConnections() {
     }
     setBorderItemsVisible(true, 0.4);
     setConnectionItemsVisible(true);
+    this->cursorMapTileRect->setSingleTileMode();
+    this->cursorMapTileRect->setVisibility(false);
 }
 
 void Editor::setDiveEmergeControls() {
@@ -348,11 +359,15 @@ void Editor::onHoveredMetatileSelectionCleared() {
 }
 
 void Editor::onSelectedMetatilesChanged() {
+    QPoint size = this->metatile_selector_item->getSelectionDimensions();
+    this->cursorMapTileRect->updateSelectionSize(size.x(), size.y());
     this->redrawCurrentMetatilesSelection();
 }
 
 void Editor::onHoveredMapMetatileChanged(int x, int y) {
-    if (x >= 0 && x < map->getWidth() && y >= 0 && y < map->getHeight()) {
+    this->playerViewRect->updateLocation(x, y);
+    this->cursorMapTileRect->updateLocation(x, y);
+    if (map_item->paintingEnabled && x >= 0 && x < map->getWidth() && y >= 0 && y < map->getHeight()) {
         int blockIndex = y * map->getWidth() + x;
         int tile = map->layout->blockdata->blocks->at(blockIndex).tile;
         this->ui->statusBar->showMessage(QString("X: %1, Y: %2, Metatile: 0x%3, Scale = %4x")
@@ -364,11 +379,17 @@ void Editor::onHoveredMapMetatileChanged(int x, int y) {
 }
 
 void Editor::onHoveredMapMetatileCleared() {
-    this->ui->statusBar->clearMessage();
+    this->playerViewRect->setVisible(false);
+    this->cursorMapTileRect->setVisible(false);
+    if (map_item->paintingEnabled) {
+        this->ui->statusBar->clearMessage();
+    }
 }
 
 void Editor::onHoveredMapMovementPermissionChanged(int x, int y) {
-    if (x >= 0 && x < map->getWidth() && y >= 0 && y < map->getHeight()) {
+    this->playerViewRect->updateLocation(x, y);
+    this->cursorMapTileRect->updateLocation(x, y);
+    if (map_item->paintingEnabled && x >= 0 && x < map->getWidth() && y >= 0 && y < map->getHeight()) {
         int blockIndex = y * map->getWidth() + x;
         uint16_t collision = map->layout->blockdata->blocks->at(blockIndex).collision;
         uint16_t elevation = map->layout->blockdata->blocks->at(blockIndex).elevation;
@@ -381,7 +402,11 @@ void Editor::onHoveredMapMovementPermissionChanged(int x, int y) {
 }
 
 void Editor::onHoveredMapMovementPermissionCleared() {
-    this->ui->statusBar->clearMessage();
+    this->playerViewRect->setVisible(false);
+    this->cursorMapTileRect->setVisible(false);
+    if (map_item->paintingEnabled) {
+        this->ui->statusBar->clearMessage();
+    }
 }
 
 QString Editor::getMovementPermissionText(uint16_t collision, uint16_t elevation){
@@ -427,7 +452,44 @@ bool Editor::setMap(QString map_name) {
     return true;
 }
 
+void Editor::onMapStartPaint(QGraphicsSceneMouseEvent *event, MapPixmapItem *item) {
+    if (!item->paintingEnabled) {
+        return;
+    }
+
+    QPointF pos = event->pos();
+    int x = static_cast<int>(pos.x()) / 16;
+    int y = static_cast<int>(pos.y()) / 16;
+    if (event->buttons() & Qt::RightButton && (map_edit_mode == "paint" || map_edit_mode == "fill")) {
+        this->cursorMapTileRect->initRightClickSelectionAnchor(x, y);
+    } else {
+        this->cursorMapTileRect->initAnchor(x, y);
+    }
+}
+
+void Editor::onMapEndPaint(QGraphicsSceneMouseEvent *event, MapPixmapItem *item) {
+    if (!item->paintingEnabled) {
+        return;
+    }
+    this->cursorMapTileRect->stopRightClickSelectionAnchor();
+    this->cursorMapTileRect->stopAnchor();
+}
+
+void Editor::setSmartPathCursorMode(QGraphicsSceneMouseEvent *event)
+{
+    bool smartPathsEnabled = event->modifiers() & Qt::ShiftModifier;
+    if (smartPathsEnabled || settings->smartPathsEnabled) {
+        this->cursorMapTileRect->setSmartPathMode();
+    } else {
+        this->cursorMapTileRect->setNormalPathMode();
+    }
+}
+
 void Editor::mouseEvent_map(QGraphicsSceneMouseEvent *event, MapPixmapItem *item) {
+    if (!item->paintingEnabled) {
+        return;
+    }
+
     if (map_edit_mode == "paint") {
         if (event->buttons() & Qt::RightButton) {
             item->updateMetatileSelection(event);
@@ -438,6 +500,7 @@ void Editor::mouseEvent_map(QGraphicsSceneMouseEvent *event, MapPixmapItem *item
                 item->floodFill(event);
             }
         } else {
+            this->setSmartPathCursorMode(event);
             item->paint(event);
         }
     } else if (map_edit_mode == "select") {
@@ -460,9 +523,24 @@ void Editor::mouseEvent_map(QGraphicsSceneMouseEvent *event, MapPixmapItem *item
     } else if (map_edit_mode == "shift") {
         item->shift(event);
     }
+
+    QPointF pos = event->pos();
+    int x = static_cast<int>(pos.x()) / 16;
+    int y = static_cast<int>(pos.y()) / 16;
+    this->playerViewRect->updateLocation(x, y);
+    this->cursorMapTileRect->updateLocation(x, y);
 }
 
 void Editor::mouseEvent_collision(QGraphicsSceneMouseEvent *event, CollisionPixmapItem *item) {
+    if (!item->paintingEnabled) {
+        return;
+    }
+
+    QPointF pos = event->pos();
+    int x = static_cast<int>(pos.x()) / 16;
+    int y = static_cast<int>(pos.y()) / 16;
+    this->playerViewRect->updateLocation(x, y);
+    this->cursorMapTileRect->updateLocation(x, y);
     if (map_edit_mode == "paint") {
         if (event->buttons() & Qt::RightButton) {
             item->updateMovementPermissionSelection(event);
@@ -503,6 +581,8 @@ void Editor::displayMap() {
     if (map_item && scene) {
         scene->removeItem(map_item);
         delete map_item;
+        scene->removeItem(this->playerViewRect);
+        scene->removeItem(this->cursorMapTileRect);
     }
 
     displayMetatileSelector();
@@ -515,6 +595,11 @@ void Editor::displayMap() {
     displayMapConnections();
     displayMapBorder();
     displayMapGrid();
+
+    this->playerViewRect->setZValue(1000);
+    this->cursorMapTileRect->setZValue(1001);
+    scene->addItem(this->playerViewRect);
+    scene->addItem(this->cursorMapTileRect);
 
     if (map_item) {
         map_item->setVisible(false);
@@ -554,6 +639,10 @@ void Editor::displayMapMetatiles() {
     map_item = new MapPixmapItem(map, this->metatile_selector_item, this->settings);
     connect(map_item, SIGNAL(mouseEvent(QGraphicsSceneMouseEvent*,MapPixmapItem*)),
             this, SLOT(mouseEvent_map(QGraphicsSceneMouseEvent*,MapPixmapItem*)));
+    connect(map_item, SIGNAL(startPaint(QGraphicsSceneMouseEvent*,MapPixmapItem*)),
+            this, SLOT(onMapStartPaint(QGraphicsSceneMouseEvent*,MapPixmapItem*)));
+    connect(map_item, SIGNAL(endPaint(QGraphicsSceneMouseEvent*,MapPixmapItem*)),
+            this, SLOT(onMapEndPaint(QGraphicsSceneMouseEvent*,MapPixmapItem*)));
     connect(map_item, SIGNAL(hoveredMapMetatileChanged(int, int)),
             this, SLOT(onHoveredMapMetatileChanged(int, int)));
     connect(map_item, SIGNAL(hoveredMapMetatileCleared()),
@@ -673,6 +762,9 @@ void Editor::displayMapEvents() {
 
 DraggablePixmapItem *Editor::addMapEvent(Event *event) {
     DraggablePixmapItem *object = new DraggablePixmapItem(event, this);
+    if (!event->usingSprite) {
+        object->setOpacity(0.7);
+    }
     events_group->addToGroup(object);
     return object;
 }
@@ -1044,6 +1136,8 @@ void DraggablePixmapItem::mouseMoveEvent(QGraphicsSceneMouseEvent *mouse) {
     if (active) {
         int x = static_cast<int>(mouse->pos().x() + this->pos().x()) / 16;
         int y = static_cast<int>(mouse->pos().y() + this->pos().y()) / 16;
+        this->editor->playerViewRect->updateLocation(x, y);
+        this->editor->cursorMapTileRect->updateLocation(x, y);
         if (x != last_x || y != last_y) {
             if (editor->selected_events->contains(this)) {
                 for (DraggablePixmapItem *item : *editor->selected_events) {
@@ -1131,6 +1225,9 @@ DraggablePixmapItem* Editor::addNewEvent(QString event_type) {
     if (project && map) {
         Event *event = Event::createNewEvent(event_type, map->name);
         event->put("map_name", map->name);
+        if (event_type == "event_heal_location") {
+            event->put("index", project->flyableMaps.length());
+        }
         map->addEvent(event);
         project->loadEventPixmaps(map->getAllEvents());
         DraggablePixmapItem *object = addMapEvent(event);