+++ src/settings.cpp
@@ -4,4 +4,6 @@ Settings::Settings()
 {
     this->smartPathsEnabled = false;
     this->betterCursors = true;
+    this->playerViewRectEnabled = false;
+    this->cursorTileRectEnabled = true;
 }