+++ CHANGELOG.md
@@ -11,15 +11,26 @@ The **"Breaking Changes"** listed below are changes that have been made in the d
 - Add "magic fill" mode to fill tool (hold down CTRL key). This fills all matching metatiles on the map, rather than only the contiguous region.
 - Add ability to import tileset palettes (JASC, .pal, .tpl, .gpl, .act).
 - Add ability to export tileset tiles as indexed .png images. The currently-selected palette is used.
+- Restore window sizes the next time the application is opened.
+- Add ability to import metatiles from Advance Map 1.92 (.bvd files).
+- Add About window that contains porymap information and changelog. (Found in file menu `Help > About Porymap`)
+- Add option to show player's in-game view when hovering the mouse on the map.
+- Add option to show an outline around the currently-hovered map tile. Its size depends on the size of the current metatile selection.
 
 ### Changed
 - Collapse the map list by default.
 - Collision view now has a transparency slider to help make it easier to view the underlying metatiles.
+- When importing tileset tiles from an image that is not indexed, the user can also provide a palette for the image. This is for the scenario where the user exports tiles and a palette from Advance Map.
+- When creating a new map, the user specifies all of the map properties in a new window prompt.
+- New maps can be created using existing layouts by right-clicking on an existing layout folder in the map list panel when sorted by "Layout".
+- The map list panel now has "expand-all" and "collapse-all" buttons.
+- Events without sprites are now partially transparent so the underlying metatile can be seen. (Warps, signs, etc.)
 
 ### Fixed
-- Fix bug where smart paths could bes auto-enabled, despite the checkbox being disabled.
+- Fix bug where smart paths could be auto-enabled, despite the checkbox being disabled.
 - Fix crash that could occur when changing the palette id in the tileset palette editor.
 - Fix crash that could occur when shrinking the number of metatiles in a tileset.
+- Fix bug where exported tile images from Advance Map were not handled correctly due to Advance Map using incorrect file extensions.
 
 ## [1.1.0] - 2018-12-27
 ### Breaking Changes